+++ libs/search-widget/src/common/label/label.utils.ts
@@ -1,14 +1,13 @@
 import type { Classification } from '@nuclia/core';
+import { getFilterFromLabel } from '@nuclia/core';
 import { isViewerOpen } from '../../core/stores/modal.store';
 import { nucliaStore } from '../../core/old-stores/main.store';
 import { typeAhead } from '../../core/stores/suggestions.store';
 
-export const labelRegexp = new RegExp(/LABEL=\{([^/]+\/[^}]+)}/, 'g');
-export const typingLabelRegexp = new RegExp(/LABEL=?\{?[^}]*}?/, 'g');
-
 export function searchBy(label: Classification) {
   isViewerOpen.set(false);
-  const labelFilter = `LABEL={${label.labelset}/${label.label}}`;
+  const labelFilter = getFilterFromLabel(label);
   typeAhead.set('');
   nucliaStore().filters.next([labelFilter]);
+  nucliaStore().triggerSearch.next();
 }