+++ libs/search-widget/src/old-components/search-input/SearchInput.svelte
@@ -1,23 +1,23 @@
 <script lang="ts">
   import { _ } from '../../core/i18n';
   import { createEventDispatcher, onMount } from 'svelte';
-  import { nucliaState, nucliaStore } from '../../core/old-stores/main.store';
+  import { nucliaState, nucliaStore, removeLabelFilter } from '../../core/old-stores/main.store';
   import { getCDN } from '../../core/utils';
   import Icon from '../../common/icons/Icon.svelte';
   import Modal from '../../common/modal/Modal.svelte';
   import Suggestions from '../suggestions/Suggestions.svelte';
   import {
     hasSuggestions,
-    suggestedIntents,
+    suggestedLabels,
     suggestedParagraphs,
     suggestionsHasError,
     typeAhead,
   } from '../../core/stores/suggestions.store';
   import { tap } from 'rxjs/operators';
   import Label from '../../common/label/Label.svelte';
-  import { map, Observable, take } from 'rxjs';
+  import { map, Observable } from 'rxjs';
   import type { Classification } from '@nuclia/core';
-  import { labelRegexp } from '../../common/label/label.utils';
+  import { getLabelFromFilter } from '@nuclia/core';
 
   export let popupSearch = false;
   export let embeddedSearch = false;
@@ -32,26 +32,19 @@
   let inputContainerElement: HTMLElement | undefined;
   let position: DOMRect | undefined;
   let showSuggestions = false;
+  let hasFilters = false;
   const filters = nucliaState().filters.pipe(
-    tap((filterQuery) => {
-      if (popupSearch) {
-        typeAhead.set(filterQuery.join(''));
+    tap((filters) => {
+      // search box size changes when there are filters or not
+      const hasFiltersNow = filters.length > 0;
+      if (hasFilters !== hasFiltersNow) {
+        hasFilters = hasFiltersNow;
+        setTimeout(() => setInputPosition());
       }
     }),
   );
   const labels: Observable<Classification[]> = filters.pipe(
-    map((filters) =>
-      filters
-        .map((filter) => {
-          const labelMatches = [...filter.matchAll(labelRegexp)];
-          if (labelMatches.length === 1) {
-            const label = labelMatches[0][1].split('/');
-            return { labelset: label[0], label: label[1] };
-          }
-          return null;
-        })
-        .filter((label) => !!label),
-    ),
+    map((filters) => filters.map((filter) => getLabelFromFilter(filter))),
   );
 
   const suggestionModalMinWidth = 384;
@@ -67,7 +60,9 @@
   });
 
   const setInputPosition = () => {
-    position = inputContainerElement?.getBoundingClientRect();
+    if (inputContainerElement) {
+      position = inputContainerElement.getBoundingClientRect();
+    }
   };
 
   const search = () => {
@@ -88,17 +83,6 @@
 
   const closeSuggestions = () => {
     showSuggestions = false;
-  };
-
-  const removeLabel = (label: Classification) => {
-    filters.pipe(take(1)).subscribe((filters) => {
-      const filterIndex = filters.findIndex((filter) => filter === `LABEL={${label.labelset}/${label.label}}`);
-      if (filterIndex > -1) {
-        const newFilters = [...filters];
-        newFilters.splice(filterIndex, 1);
-        nucliaStore().filters.next(newFilters);
-      }
-    });
   };
 </script>
 
@@ -150,13 +134,13 @@
     {/if}
   </div>
 
-  {#if $filters.length > 0 && (embeddedSearch || searchBarWidget)}
+  {#if $filters.length > 0}
     <div class="filters-container">
       {#each $labels as label (label.label)}
         <Label
           {label}
           removable
-          on:remove={() => removeLabel(label)} />
+          on:remove={() => removeLabelFilter(label)} />
       {/each}
     </div>
   {/if}
@@ -170,7 +154,7 @@
   <div class="sw-suggestions-container">
     <Suggestions
       paragraphs={$suggestedParagraphs}
-      intents={$suggestedIntents} />
+      labels={$suggestedLabels} />
   </div>
 </Modal>